# Add performance test and minor code refactor