--- a/.gitignore
+++ b/.gitignore
@@ -10,6 +10,8 @@ lerna-debug.log*
 node_modules
 *.local
 
+tests/out
+
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json