… dist/index.d.ts
… dist/index.js