@Cainier 2024-05-19
Thank you very much for your contribution to this project:

1. NPM registry error is because I mistakenly modified the global npm private package when developing another project. Thanks for finding and raising the issue, I will delete package-lock.json and re-execute npm i to generate the file

2. Model `ft:gpt-3.5-turbo-1106:opensftp::8IWeqPit` only works when testing with my own accesskey. This is indeed a problem, I will replace it with environment variables as per the scheme in your submission

3. Regarding WASM, Starting from v1.1 version, changed from [@dqbd/tiktoken](https://github.com/dqbd/tiktoken) to js-tiktoken, because using wasm on the web requires configuration of vite/webpack https://github.com/Cainier/gpt-tokens/tree/v1.0.9

If have the need to use wasm, I will try add `useWASM` configuration in next version

I will try to use CI/CD pipeline for automated testing and building, But accessKey is a problem, maybe use GitHub environment variables can fix it